MB-22819. Wrong int type for CAS and expiration.
Reviewed-on: http://review.couchbase.org/73707
Tested-by: Gerald Sangudi <gerald@couchbase.com>

+++ datastore/couchbase/couchbase.go
@@ -849,8 +849,8 @@ func (b *keyspace) Fetch(keys []string, context datastore.QueryContext) ([]value
 				"id":         k,
 				"cas":        v.Cas,
 				"type":       meta_type,
-				"flags":      int64(flags),
-				"expiration": int64(expiration),
+				"flags":      flags,
+				"expiration": expiration,
 			})
 
 			// Uncomment when needed
